+++ R/rice.R
@@ -1,9 +1,11 @@
 # to do: 
 
+# write a decontaminate function (e.g. to find the contamination required to explain away a Shroud date)
+
 # 'combine' to combine multiple distributions into 1 multiplied distribution
 
 # soil F14C simulation from Vegard
 
 # terr-marine contribution calculation
 
-# AMS background calculations and fractionation corrections
+# AMS background and fractionation corrections